+++ src/components/Card.tsx
@@ -1,7 +1,7 @@
-export const Card = (props) => {
+export const Card = ({ children }) => {
   return (
     <div className="rounded-2xl p-4 shadow-md">
-      {props.children}
+      {children}
     </div>
   )
 }
